Use parent ListStore instead of its flat index
The flat indices into the TreeListModel are ephemereal, as they change
when parts of the tree are expanded or collapsed.
This means we can't use them for undo/redo operations, as they might
no longer be valid (if e.g. a group was expanded in the meantime).
Instead, we keep track of the parent ListStore and index within it.